adapters/mcp-server/pyproject.toml
[project]
name = "aura-mcp-server"
version = "0.1.0"
description = "MCP Server for Aura Platform"
authors = [
    {name = "Aura Platform", email = "platform@aura.ai"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [ # MCP SDK
    "httpx>=0.27.0", # Async HTTP client
    "pynacl>=1.5.0", # Ed25519 signing
    "python-dotenv>=1.2.1",
    "fastapi>=0.128.0", # If needed for MCP # ASGI server
    "uvicorn>=0.40.0",
    "fastmcp>=3.0.0b1",
]

[tool.ruff]
target-version = "py313"
line-length = 88
exclude = [
    "**/proto/**",  # Auto-generated protobuf files
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]
